chore: 🤖 remove /src/plugins/__mocks__ folder
